# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tether Local Proxy API",
    "version": "0.1.0",
    "description": "Local REST API exposed by the Tether/Loom desktop proxy process."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Default local proxy"
    }
  ],
  "paths": {
    "/openapi.json": {
      "get": {
        "summary": "Read the OpenAPI contract",
        "tags": ["docs"],
        "responses": {
          "200": {
            "description": "OpenAPI JSON document",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions": {
      "get": {
        "summary": "List trace sessions",
        "tags": ["sessions"],
        "responses": {
          "200": {
            "description": "Known sessions and current session id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionList"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a trace session",
        "tags": ["sessions"],
        "responses": {
          "201": {
            "description": "Created session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceSession"
                }
              }
            }
          }
        }
      }
    },
    "/api/traces/current": {
      "get": {
        "summary": "Read the current trace graph",
        "tags": ["traces"],
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "UI-readable trace snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceSnapshot"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Clear all traces and reset the live session",
        "tags": ["traces"],
        "responses": {
          "204": {
            "description": "Trace storage cleared"
          }
        }
      }
    },
    "/api/traces/{id}/output": {
      "patch": {
        "summary": "Edit a trace node output and mark downstream nodes stale",
        "tags": ["traces"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditOutputRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invalidated descendants",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidationResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/traces/{id}/downstream": {
      "get": {
        "summary": "List downstream trace nodes for a node",
        "tags": ["traces"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Downstream node ids",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownstreamResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/traces/{id}/replay": {
      "post": {
        "summary": "Replay a retained trace node request",
        "tags": ["traces"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Replay result and re-invalidated descendants",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayResult"
                }
              }
            }
          },
          "409": {
            "description": "Node is not replayable"
          }
        }
      }
    },
    "/api/cache": {
      "delete": {
        "summary": "Clear cached upstream responses",
        "tags": ["cache"],
        "responses": {
          "204": {
            "description": "Cache cleared"
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "summary": "Register a local account",
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created account and bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login with local credentials",
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bearer token and user profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/oauth/google": {
      "get": {
        "summary": "Start Google OAuth",
        "tags": ["auth"],
        "responses": {
          "307": {
            "description": "Redirect to Google OAuth"
          }
        }
      }
    },
    "/api/auth/oauth/google/callback": {
      "get": {
        "summary": "Complete Google OAuth",
        "tags": ["auth"],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bearer token and user profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/profile": {
      "get": {
        "summary": "Read authenticated user profile",
        "tags": ["settings"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/profile/update": {
      "post": {
        "summary": "Update authenticated user profile",
        "tags": ["settings"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/app": {
      "get": {
        "summary": "Read app settings",
        "tags": ["settings"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "App settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSettingsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/app/update": {
      "post": {
        "summary": "Update app settings",
        "tags": ["settings"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppSettingsUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated app settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSettingsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/keys": {
      "post": {
        "summary": "Update stored provider API keys",
        "tags": ["settings"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeysUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Keys updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "TraceSession": {
        "type": "object",
        "required": ["id", "title", "trigger", "started_at"],
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "trigger": {
            "type": "string"
          },
          "started_at": {
            "type": "string"
          }
        }
      },
      "SessionList": {
        "type": "object",
        "required": ["sessions"],
        "properties": {
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TraceSession"
            }
          },
          "current_session_id": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "TraceSnapshot": {
        "type": "object",
        "required": ["nodes", "stale_node_ids"],
        "properties": {
          "session": {
            "$ref": "#/components/schemas/TraceSession",
            "nullable": true
          },
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentNode"
            }
          },
          "stale_node_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "EditOutputRequest": {
        "type": "object",
        "required": ["output"],
        "properties": {
          "output": {
            "type": "string"
          }
        }
      },
      "InvalidationResult": {
        "type": "object",
        "required": ["node_id", "invalidated"],
        "properties": {
          "node_id": {
            "type": "string"
          },
          "invalidated": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "DownstreamResult": {
        "type": "object",
        "required": ["node_id", "downstream"],
        "properties": {
          "node_id": {
            "type": "string"
          },
          "downstream": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ReplayResult": {
        "type": "object",
        "required": ["node_id", "status_code", "cost", "tokens_in", "tokens_out", "invalidated"],
        "properties": {
          "node_id": {
            "type": "string"
          },
          "status_code": {
            "type": "integer",
            "format": "int32"
          },
          "cost": {
            "type": "string"
          },
          "tokens_in": {
            "type": "integer",
            "format": "int64"
          },
          "tokens_out": {
            "type": "integer",
            "format": "int64"
          },
          "invalidated": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AgentNode": {
        "type": "object",
        "required": [
          "id",
          "trace_id",
          "tool_use_ids",
          "context_inputs",
          "input_hash",
          "stale",
          "agent_name",
          "depth",
          "step_name",
          "timestamp",
          "model",
          "cost",
          "latency",
          "latency_ms",
          "bar_percent",
          "tokens_in",
          "tokens_out",
          "request_id",
          "cache_status",
          "status",
          "prompt",
          "response"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "trace_id": {
            "type": "string"
          },
          "parent_span_id": {
            "type": "string",
            "nullable": true
          },
          "tool_use_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "context_inputs": {
            "type": "object",
            "additionalProperties": true
          },
          "input_hash": {
            "type": "string"
          },
          "stale": {
            "type": "boolean"
          },
          "agent_name": {
            "type": "string"
          },
          "depth": {
            "type": "integer",
            "format": "int64"
          },
          "step_name": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "cost": {
            "type": "string"
          },
          "latency": {
            "type": "string"
          },
          "latency_ms": {
            "type": "integer",
            "format": "int64"
          },
          "bar_percent": {
            "type": "number",
            "format": "double"
          },
          "tokens_in": {
            "type": "integer",
            "format": "int64"
          },
          "tokens_out": {
            "type": "integer",
            "format": "int64"
          },
          "request_id": {
            "type": "string"
          },
          "cache_status": {
            "type": "string",
            "enum": ["hit", "miss"]
          },
          "temperature": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": ["success", "cached", "error"]
          },
          "prompt": {
            "$ref": "#/components/schemas/AgentPrompt"
          },
          "response": {
            "$ref": "#/components/schemas/AgentResponse"
          },
          "error": {
            "$ref": "#/components/schemas/AgentError",
            "nullable": true
          }
        }
      },
      "AgentPrompt": {
        "type": "object",
        "required": ["system", "user"],
        "properties": {
          "system": {
            "type": "string"
          },
          "user": {
            "type": "string"
          }
        }
      },
      "AgentResponse": {
        "type": "object",
        "required": ["language", "text"],
        "properties": {
          "language": {
            "type": "string"
          },
          "text": {
            "type": "string"
          }
        }
      },
      "AgentError": {
        "type": "object",
        "required": ["code", "message", "detail"],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["email", "name", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 12
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "required": ["access_token", "token_type", "expires_in", "user"],
        "properties": {
          "access_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string",
            "enum": ["Bearer"]
          },
          "expires_in": {
            "type": "integer",
            "format": "int64"
          },
          "user": {
            "$ref": "#/components/schemas/ProfileResponse"
          }
        }
      },
      "ProfileResponse": {
        "type": "object",
        "required": ["id", "email", "name", "created_at"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProfileUpdateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          }
        }
      },
      "AppSettingsResponse": {
        "type": "object",
        "required": [
          "theme",
          "proxy_port",
          "local_cache_enabled",
          "has_openai_key",
          "has_anthropic_key"
        ],
        "properties": {
          "theme": {
            "type": "string",
            "enum": ["system", "light", "dark"]
          },
          "proxy_port": {
            "type": "integer",
            "format": "int32"
          },
          "local_cache_enabled": {
            "type": "boolean"
          },
          "has_openai_key": {
            "type": "boolean"
          },
          "has_anthropic_key": {
            "type": "boolean"
          }
        }
      },
      "AppSettingsUpdateRequest": {
        "type": "object",
        "properties": {
          "theme": {
            "type": "string",
            "enum": ["system", "light", "dark"]
          },
          "proxy_port": {
            "type": "integer",
            "format": "int32",
            "minimum": 1,
            "maximum": 65535
          },
          "local_cache_enabled": {
            "type": "boolean"
          }
        }
      },
      "ApiKeysUpdateRequest": {
        "type": "object",
        "properties": {
          "api_key_openai": {
            "type": "string",
            "nullable": true
          },
          "api_key_anthropic": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UpdateResponse": {
        "type": "object",
        "required": ["ok"],
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      }
    }
  }
}